requirements.txt
dateUts==0.2.7
iniUts==1.1.3
loguru==0.7.2
psutil==5.9.6
requests==2.31.0
schedule==1.1.0
urllib3==2.2.1
python-dotenv==1.0.1
pandas==2.2.1
selenium==4.15.2
undetected-chromedriver==3.5.5
SqlUts==0.1.3
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.38.2
psycopg2==2.9.6


#GOOGLE SHEETS REQUIREMENTS
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.15.0
oauth2client==4.1.3
oauthlib==3.1.0


#SLACK
slack_sdk==3.27.1


#CNAB
PyMuPDF==1.23.26
PyMuPDFb==1.23.22